+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Polyverse Boost Automatic Application Modernization
 - N/A
 
 ### Enhancements
+- Project Analysis Summary and File Analysis Output are automatically generated during UI analysis runs
 - Improvements to readability, style and source line numbers in HTML and PDF output
 - Enabled Right-click in File Explorer for showing PDF, HTML and Markdown of any Boost Notebook
 - All Output files are stored in Boost Output folder - for ease of packaging and publishing